Refactor

+++ src/components/transform.ts
@@ -1,6 +1,6 @@
 import { defineComponent, Types } from "bitecs";
 
 export const Transform = defineComponent({
-  x: Types.i32,
-  y: Types.i32,
+  x: Types.f32,
+  y: Types.f32,
 });
